+++ packages/contracts/contracts/test/unit/BalanceManager.t.sol
@@ -9,7 +9,7 @@ import "@openzeppelin/contracts/utils/Strings.sol";
 
 import {IJoinSplitVerifier} from "../../interfaces/IJoinSplitVerifier.sol";
 import {ISubtreeUpdateVerifier} from "../../interfaces/ISubtreeUpdateVerifier.sol";
-import {OffchainMerkleTree, OffchainMerkleTreeData} from "../../libs/OffchainMerkleTree.sol";
+import {LibOffchainMerkleTree, OffchainMerkleTree} from "../../libs/OffchainMerkleTree.sol";
 import {TestJoinSplitVerifier} from "../harnesses/TestJoinSplitVerifier.sol";
 import {TestSubtreeUpdateVerifier} from "../harnesses/TestSubtreeUpdateVerifier.sol";
 import {OperationUtils} from "../../libs/OperationUtils.sol";
@@ -24,7 +24,7 @@ import {AssetUtils} from "../../libs/AssetUtils.sol";
 import "../../libs/Types.sol";
 
 contract BalanceManagerTest is Test {
-    using OffchainMerkleTree for OffchainMerkleTreeData;
+    using LibOffchainMerkleTree for OffchainMerkleTree;
     using stdJson for string;
     using OperationLib for Operation;